LayoutTests/fast/forms/textarea-placeholder-dom-property.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=21248">https://bugs.webkit.org/show_bug.cgi?id=21248</a>');

var textarea = document.createElement('textarea');

shouldBeEqualToString('textarea.placeholder', '');

textarea.setAttribute('placeholder', 'p1')
shouldBeEqualToString('textarea.placeholder', 'p1');
shouldBeEqualToString('textarea.value', '');

textarea.placeholder = 'p2';
shouldBeEqualToString('textarea.getAttribute("placeholder")', 'p2');
shouldBeEqualToString('textarea.placeholder', 'p2');
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
